Replaced awful while loops with asynchronous callbacks, and replaced image loader with a nicer cached asynchronous implementation
